converted intigers to double. fixed common known issues.

+++ lib/outer_pages/login.dart
@@ -21,7 +21,7 @@ class _LoginPageState extends State<LoginPage> {
   String _email = '';
   String _password = '';
   final String _profile_picture = '';
-  int _wallet_balance = 0;
+  double _wallet_balance = 0.0; // Change type to double
   String _selectedCountry = '';
   bool _isLoadingLocation = true;
   //add phone registering later
@@ -354,9 +354,9 @@ class _LoginPageState extends State<LoginPage> {
           'username': _username,
           'email': _email,
           'profile_picture': '', // Empty profile picture field
-          'wallet_balance': 0, // Initial wallet balance of 0
+          'wallet_balance': 0.0, // Initial wallet balance as double
           'location': _selectedCountry, // Save the selected country
-          'total_points': 0, // Initial total points of 0
+          'total_points': 0.0, // Initial total points as double
           'level': 1, // Initial level of 0
         });
